ui-kit: delete 945 LOC of IDE-only @Preview composables (#1449)
The 23 @Preview composables in shared/ui-kit/src/main/.../preview/ (Components/
Primitives/ThemePreview) are IDE-only, ship in the release APK as dead weight
(isMinifyEnabled=false, so R8 doesn't strip them), and are redundant with the
sanctioned DesignRenders (#555) JVM render path that already replaced the
Android-Studio preview workflow (D22 hard-cut). Delete all three files, port
the one coverage gap — the design-token/hex swatch sheet — into DesignRenders
as ThemeSwatchRenderFixtures + a designTokenSwatches() @Test, and drop the now-
unused compose.ui.tooling.preview dep. grep shows no @Preview left in ui-kit
main; render PNG inspected; :shared:ui-kit:testDebugUnitTest green.

Closes #1449
